Tier 0 review R37 (3 rounds): fix loop/budget/natives bugs
Round 1 (6 bugs):
- loop: budget.spend() return ignored → turn Completed even when breached;
  now fails with resource/budget-exhausted on rejection
- budget: deriveChild didn't guard negative remaining → could inflate budget;
  now clamps reserve ≥ 0
- budget: exhausted() used `total` but spend() rejects at `abortThreshold`
  → mismatch; now exhausted = spent ≥ abortThreshold (+ unlimited exemption)
- natives: u128 not napi-compatible → switched to u64 (584-year range)
- types: added missing §4 types TurnContext + ExtensionAPI
- loop: pre-abort guard (don't emit start after end)

Round 2 (3 bugs):
- loop: turn/start event LOST — IIFE emitted synchronously before caller
  could subscribe (handle.on). Now deferred to queueMicrotask so callers
  subscribe first. Verified: start event now visible in print + SDK.
- loop: added pre-stream budget gate (§4: check exhausted() BEFORE streaming)
- budget: freeBudget(total=0,unlimited) falsely exhausted → exhausted() now
  exempts unlimited budgets

Round 3: 11/11 edge-case tests pass (budget CAS, deriveChild clamp, cancel
terminal, no-profile failure); tsc + clippy -D warnings clean.

+++ crates/natives/src/lib.rs
@@ -56,19 +56,24 @@ pub fn grep(_pattern: String, _root: String) -> Result<Vec<GrepHit>> {
 /// Monotonic has no epoch; returns nanos since a process-local anchor.
 /// TS core.time mirrors this; never call Instant elsewhere in Rust.
 #[napi]
-pub fn now_monotonic_nanos() -> Result<u128> {
+pub fn now_monotonic_nanos() -> Result<u64> {
   use std::sync::OnceLock;
   use std::time::Instant;
   static ANCHOR: OnceLock<Instant> = OnceLock::new();
   let anchor = ANCHOR.get_or_init(Instant::now);
-  Ok(anchor.elapsed().as_nanos())
+  Ok(anchor.elapsed().as_nanos() as u64)
 }
 
 /// Wallclock nanos since UNIX_EPOCH — the ONLY wall source (invariant #10).
 #[napi]
-pub fn now_wallclock_nanos() -> Result<u128> {
+pub fn now_wallclock_nanos() -> Result<u64> {
   use std::time::{SystemTime, UNIX_EPOCH};
-  Ok(SystemTime::now().duration_since(UNIX_EPOCH).unwrap_or_default().as_nanos())
+  Ok(
+    SystemTime::now()
+      .duration_since(UNIX_EPOCH)
+      .unwrap_or_default()
+      .as_nanos() as u64,
+  )
 }
 
 /// ABI stamp — refuses mismatched-release binaries (R25-23).
